refactor: merge JdbcTemplate to SelectJdbcTemplate
