+++ slipp/src/main/java/slipp/dao/JdbcTemplate.java
@@ -6,7 +6,10 @@ import slipp.support.db.ConnectionManager;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 
 public abstract class JdbcTemplate {
     private static final Logger logger = LoggerFactory.getLogger(JdbcTemplate.class);
@@ -23,5 +26,43 @@ public abstract class JdbcTemplate {
         }
     }
 
+    List query(String sql) {
+        try (Connection con = ConnectionManager.getConnection();
+             PreparedStatement pstmt = con.prepareStatement(sql)) {
+            setValues(pstmt);
+
+            ResultSet resultSet = pstmt.executeQuery();
+            List<Object> objects = new ArrayList<>();
+            while (resultSet.next()) {
+                objects.add(mapRow(resultSet));
+            }
+            return objects;
+        } catch (SQLException e) {
+            logger.error(e.getMessage(), e);
+            throw new RuntimeException();
+        }
+    }
+
+    Object queryForObject(String sql) {
+        try (Connection con = ConnectionManager.getConnection();
+             PreparedStatement pstmt = con.prepareStatement(sql)) {
+            setValues(pstmt);
+
+            ResultSet resultSet = pstmt.executeQuery();
+            List<Object> objects = new ArrayList<>();
+            while (resultSet.next()) {
+                objects.add(mapRow(resultSet));
+            }
+            logger.debug("objects size : {} ", objects.size());
+            return objects.get(0);
+        } catch (SQLException e) {
+            logger.error(e.getMessage(), e);
+            throw new RuntimeException();
+        }
+    }
+
     abstract void setValues(PreparedStatement pstmt) throws SQLException;
+
+    abstract Object mapRow(ResultSet resultSet) throws SQLException;
+
 }
